business loan exceptions done

+++ src/main/java/com/revature/RevPayBackend/service/BusinessLoanServiceImpl.java
@@ -1,6 +1,7 @@
 package com.revature.RevPayBackend.service;
 
 import com.revature.RevPayBackend.entity.BusinessLoan;
+import com.revature.RevPayBackend.exceptions.UserExceptions.IdNotFoundException;
 import com.revature.RevPayBackend.repository.BusinessLoanRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +21,8 @@ public class BusinessLoanServiceImpl implements BusinessLoanService {
     }
 
     @Override
-    public BusinessLoan getById(Long businessLoanId) {
+    public BusinessLoan getById(Long businessLoanId) throws IdNotFoundException {
+        if(!businessLoanRepository.existsById(businessLoanId)) throw new IdNotFoundException();
         return businessLoanRepository.findById(businessLoanId).get();
     }
 
@@ -30,12 +32,14 @@ public class BusinessLoanServiceImpl implements BusinessLoanService {
     }
 
     @Override
-    public BusinessLoan update(BusinessLoan businessLoan) {
+    public BusinessLoan update(BusinessLoan businessLoan) throws IdNotFoundException {
+        if(!businessLoanRepository.existsById(businessLoan.getLoanId())) throw new IdNotFoundException();
         return businessLoanRepository.save(businessLoan);
     }
 
     @Override
-    public boolean delete(Long businessLoanId) {
+    public boolean delete(Long businessLoanId) throws IdNotFoundException {
+        if(!businessLoanRepository.existsById(businessLoanId)) throw new IdNotFoundException();
         boolean found = businessLoanRepository.existsById(businessLoanId);
         if(found) {
             businessLoanRepository.deleteById(businessLoanId);
